requirements.txt
flask
pymongo[srv]
python-docx
requests
gunicorn
huggingface_hub